requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4
langchain-text-splitters
langchain-ollama
langchain-huggingface
sentence-transformers
faiss-cpu
chromadb
streamlit